--warning
clang complained about putting a pointer into a format that expects an
int. Avoid that by casting the pointer to a ptrdiff_t which is an int type.

+++ hooklib/variabledumping.h
@@ -164,7 +164,7 @@ TRACELIB_SPECIALIZE_CONVERSION_USING_SSTREAM(std::string)
 template <>
 inline VariableValue convertVariable( const void *val ) {
     std::vector<char> buf( 11 );
-    snprintf( &buf[0], 11, "0x%08X", val );
+    snprintf( &buf[0], 11, "0x%08X", (ptrdiff_t)val );
     return convertVariable( &buf[0] );
 }
 #if defined(_MSC_VER)
